Reorganize tests
Move test in `env` module to integration tests
and split off error-focused tests.
